medctrl-backend/requirements.txt
django>=4.0,<4.1
djangorestframework==3.13.1
django-cors-headers
mysqlclient
black
pylint
uvicorn
gunicorn
django-rest-knox